+++ meerkat-dbm/src/dbm/__test__/dbm-parallel.spec.ts
@@ -3,7 +3,8 @@ import { DBMLogger } from '../../logger/logger-types';
 import { BROWSER_RUNNER_TYPE } from '../../window-communication/runner-types';
 import { DBMParallel } from '../dbm-parallel/dbm-parallel';
 import { IFrameRunnerManager } from '../dbm-parallel/runner-manager';
-import { InstanceManager, MockFileManager } from './mock';
+import { MockFileManager } from './dbm.spec';
+import { InstanceManager } from './mock';
 
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 //@ts-ignore
